Rework task util
